Melhorias e separacao de arquivos de conversao
