+++ src/store.js
@@ -1,48 +1,9 @@
 import Vue from "vue";
 import Vuex from "vuex";
+import conversions from "./filters/conversions";
 
 Vue.use(Vuex);
 
-
-function setWindDirection(val) {
-    if (val >= 0 && val <= 11.25) { return 'N'; }
-    if (val > 11.25 && val <= 33.75) { return 'NNE'; }
-    if (val > 33.75 && val <= 56.25) { return 'NE'; }
-    if (val > 56.25 && val <= 78.75) { return 'LNE'; }
-    if (val > 78.75 && val <= 101.25) { return 'L'; }
-    if (val > 101.25 && val <= 123.75) { return 'LSE'; }
-    if (val > 123.75 && val <= 146.25) { return 'SE'; }
-    if (val > 146.25 && val <= 168.75) { return 'SSE'; }
-    if (val > 168.75 && val <= 191.25) { return 'S'; }
-    if (val > 191.25 && val <= 213.75) { return 'SSO'; }
-    if (val > 213.75 && val <= 236.25) { return 'SO'; }
-    if (val > 236.25 && val <= 258.75) { return 'OSE'; }
-    if (val > 258.75 && val <= 281.25) { return 'O'; }
-    if (val > 281.25 && val <= 303.75) { return 'ONO'; }
-    if (val > 303.75 && val <= 326.25) { return 'NO'; }
-    if (val > 326.25 && val <= 348.75) { return 'NNO'; }
-}
-
-function setToCelsius(val) {
-    var temperature = (val - 32) * 5 / 9;
-    return Math.round(temperature);
-}
-
-function setMedTemperature(low, high) {
-    return ((low + high)) / 2;
-}
-
-function setPressure(val) {
-    return Math.round(val * 33.863886666667);
-}
-
-function setWindSpeed(val) {
-    return Math.round(val * 3.6);
-}
-
-function setRoundVal(val) {
-    return Math.round(val);
-}
 
 
 export default new Vuex.Store({
@@ -71,16 +32,16 @@ export default new Vuex.Store({
             state.weather.city = weather.location.city;
             state.weather.region = weather.location.region;
             state.weather.condition = weather.current_observation.condition.text;
-            state.weather.today = `${setRoundVal(weather.current_observation.condition.temperature)}`;
-            state.weather.tomorrow = `${setRoundVal(setMedTemperature(weather.forecasts[1].high , weather.forecasts[1].low))}`;
-            state.weather.after = `${setRoundVal(setMedTemperature(weather.forecasts[2].high , weather.forecasts[2].low))}`;
-            state.weather.todayC = `${setToCelsius(weather.current_observation.condition.temperature)}`;
-            state.weather.tomorrowC = `${setToCelsius(setMedTemperature(weather.forecasts[1].high , weather.forecasts[1].low))}`;
-            state.weather.afterC = `${setToCelsius(setMedTemperature(weather.forecasts[2].high , weather.forecasts[2].low))}`;
-            state.weather.windSpeed = `${setWindSpeed(weather.current_observation.wind.speed)}`;
-            state.weather.windDirection = `${setWindDirection(weather.current_observation.wind.direction)}`;
+            state.weather.today = `${conversions.setRoundVal(weather.current_observation.condition.temperature)}`;
+            state.weather.tomorrow = `${conversions.setRoundVal(conversions.setMedTemperature(weather.forecasts[1].high , weather.forecasts[1].low))}`;
+            state.weather.after = `${conversions.setRoundVal(conversions.setMedTemperature(weather.forecasts[2].high , weather.forecasts[2].low))}`;
+            state.weather.todayC = `${conversions.setToCelsius(weather.current_observation.condition.temperature)}`;
+            state.weather.tomorrowC = `${conversions.setToCelsius(conversions.setMedTemperature(weather.forecasts[1].high , weather.forecasts[1].low))}`;
+            state.weather.afterC = `${conversions.setToCelsius(conversions.setMedTemperature(weather.forecasts[2].high , weather.forecasts[2].low))}`;
+            state.weather.windSpeed = `${conversions.setWindSpeed(weather.current_observation.wind.speed)}`;
+            state.weather.windDirection = `${conversions.setWindDirection(weather.current_observation.wind.direction)}`;
             state.weather.humidity = weather.current_observation.atmosphere.humidity;
-            state.weather.pressure = `${setPressure(weather.current_observation.atmosphere.pressure)}`;
+            state.weather.pressure = `${conversions.setPressure(weather.current_observation.atmosphere.pressure)}`;
         }
     }
 });
